+++ src/ui/dialog/loader.rs
@@ -74,7 +74,7 @@ impl Component for LoaderPopup {
                     format!("{} message", self.operation_name),
                     output,
                 ))),
-                AppAction::RefreshTab,
+                AppAction::MarkTabsStale,
             ]),
             Ok(_) => AppAction::PopupDone,
             Err(err) => AppAction::SetPopup(Box::new(MessagePopup::new(